nn, gemma, qwen: hold the key-value cache in fp16 rather than float32
The cache rounded every entry through fp16 — llama.cpp's default cache type,
and what the reference stores — and then kept the result in a float32. Four
bytes carrying two bytes of information, written once and read on every token
afterwards.

Two kernels let it hold the sixteen bits it has: DotF32Half for the scores and
AxpyHalf for the mix, both widening with VCVTPH2PS. Both mirror the float32
kernels instruction for instruction, including the fused multiply-add in the
mix and the unfused pair in its tail. That is not a style choice: a mixture of
experts routes on a hard top-k of its router's logits, so summing the same
products in a different order picks a different expert and the answer diverges
for real. A scalar version of the same arithmetic failed
TestMoEGreedyMatchesTheReference at the third token.

The measurement, on the 12B at four thousand positions of context, where the
cache is about a tenth of what a token reads:

| | ms/token | tokens/s |
|---|---|---|
| float32 cache | 231.6 | 4.32 |
| fp16 cache | 213.4 | 4.69 |

×1.08, and more than the traffic alone predicts — half the cache also fits the
levels above it better. The cache's footprint halves exactly, which is what a
server holding several conversations at once actually runs out of.

BenchmarkForwardAtDepth is new: BenchmarkForward fills thirty-two positions and
measures what reading the weights costs, which is the wrong depth to see a cache
at all.

+++ gemma/attention_test.go
@@ -142,7 +142,7 @@ func TestSharingBlockWritesNothing(t *testing.T) {
 	cache, s := NewCache(cfg), NewScratch(cfg)
 	replayBlock(t, f, cfg, w, cache, s, 13, false)
 
-	before := append([]float32(nil), cache.Layers[13].K...)
+	before := append([]uint16(nil), cache.Layers[13].K...)
 	replayBlock(t, f, cfg, w, cache, s, 15, false)
 
 	for i := range before {
